Moved datatypes into seperate files
